Positional Tracking & other stuff..

+++ Projects/Android/jni/quake2/src/game/q_shared.h
@@ -488,12 +488,12 @@ typedef struct
 {
 	pmtype_t	pm_type;
 
-	short		origin[3];		// 12.3
-	short		velocity[3];	// 12.3
+	float		origin[3];		// 12.3
+	float		velocity[3];	// 12.3
 	byte		pm_flags;		// ducked, jump_held, etc
 	byte		pm_time;		// each unit = 8 ms
-	short		gravity;
-	short		delta_angles[3];	// add to command angles to get view direction
+	float		gravity;
+	float		delta_angles[3];	// add to command angles to get view direction
 									// changed by spawns, rotating objects, and teleporters
 } pmove_state_t;
 
@@ -512,7 +512,7 @@ typedef struct usercmd_s
 	byte	msec;
 	byte	buttons;
 	short	angles[3];
-	short	forwardmove, sidemove, upmove;
+	float	forwardmove, sidemove, upmove;
 	byte	impulse;		// remove?
 	byte	lightlevel;		// light level the player is standing on
 } usercmd_t;
